Move compat code into compat package, reformat code
