+++ core/src/main/scala-2.12-/cats/collections/compat/HashSetCompatCompanion.scala
@@ -20,6 +20,7 @@
  */
 
 package cats.collections
+package compat
 
 import scala.collection.immutable.Set
 import scala.collection.AbstractSet
