+++ core/src/main/scala-2.13+/cats/collections/compat/HashSetCompatCompanion.scala
@@ -20,6 +20,7 @@
  */
 
 package cats.collections
+package compat
 
 import scala.collection.immutable.AbstractSet
