+++ tests/src/test/scala/cats/collections/DefaultScalaCheckPropertyCheckConfig.scala
@@ -27,6 +27,7 @@ object DefaultScalaCheckPropertyCheckConfig {
   final val default: Parameters = Parameters.default
     .withMinSuccessfulTests(if (BuildInfo.isJvm) 50 else 5)
     .withMaxDiscardRatio(if (BuildInfo.isJvm) 5 else 50)
+    .withMaxSize(if (BuildInfo.isJvm) 10 else 5)
     .withMinSize(0)
     .withWorkers(if (BuildInfo.isJvm) 2 else 1)
 }
